sst-bot/pyproject.toml
[tool.poetry]
name = "sst-bot"
version = "0.1.0"
description = ""
authors = ["Martin <martin.jarnier@epita.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
pandas = "^2.1.4"
transformers = "^4.36.2"
torch = "^2.1.2"
torchaudio = "^2.1.2"
scikit-learn = "^1.4.1.post1"
vadersentiment = "^3.3.2"
spacy = "^3.7.4"
gensim = "^4.3.2"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
pyannote-audio = {git = "https://github.com/pyannote/pyannote-audio"}
ipython = "^8.22.1"
pydub = "^0.25.1"
webrtcvad = "^2.0.10"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
